Corrected linters
