+++ src/checkbox.js
@@ -1,20 +1,4 @@
-var tasksArray = [
-  {
-    description: 'Thinking of a fat joint',
-    completed: false,
-    index: 0,
-  },
-  {
-    description: 'Rolling a fat joint',
-    completed: false,
-    index: 1,
-  },
-  {
-    description: 'Smoking of a fat joint',
-    completed: false,
-    index: 2,
-  },
-];
+import {tasksArray} from './index.js'
 
 const tasks = document.querySelector('.tasks');
 
@@ -40,23 +24,8 @@ function updateStorage() {
   localStorage.setItem('tasksArray', JSON.stringify(tasksArray));
 }
 
-function displayCheckbox() {
-  const checkbox = document.querySelectorAll('.checkbox');
-  const inputDisplay = document.querySelectorAll('.inputDisplay');
-  if (localStorage.getItem('tasksArray')) {
-    tasksArray = JSON.parse(localStorage.getItem('tasksArray'));
-    for (let i = 0; i < checkbox.length; i += 1) {
-      if (tasksArray[i].completed === true) {
-        checkbox[i].checked = true;
-        inputDisplay[i].classList.add('checkedDisplay');
-      }
-    }
-  }
-}
-
 export {
   updateStorage,
   render,
   tasksArray,
-  displayCheckbox,
 };
